Adds custom spack extensions for OLCF tasks.
A custom spack extension for OLCF-related tasks is now available as a spack
subcommand when added to a spack configurations's extensions key:

```
config:
  extensions:
    - $git/share/spack/extensions/spack-olcf
```

Currently, only a utility for displaying all of the concretized specs in an
environment's lock file is added. This allows admins to review the specs in a
human-readable way without having to reconcretize the entire environment. It
also shows additional information about the packages in the environment
(namespaces and dependency types) beyond the default `spack concretize -f`.