Adds custom OLCF commands to Spack CLI.
The custom OLCF command adds initially a utility to list the concretized package
list for an environment with additional verbosity needed to verify packages are
concretized as intended.

It also adds a deploy subcommand that mimics the standard spack environment
deploy, but uses the verbose spec listing described above and will ultimately
show the verbose concretized spec for an explicit spec just prior to starting
the build.