Removes outdated configurations for systems other than Peak
The plan moving forward is to use only spack environments for each system. The
older config files such as `<host>/etc/spack/compilers.yaml` are obsolete and
outdated.

Of the files removed in this commit, only `summit/etc/spack.yaml` is likely to
be somewhat relevent. However, the organizational structure of the environments
has been changed (using the work on Peak as a template). Therefore, the contents
and location of the removed summit environment file are not in-line with what
the final goal and it will be easier to just clone the peak environment for
summit in a later commit.