Snapshots static modulefiles for Peak.
The structure of static modulefiles is changed to eliminate the need for
relative symlinks from the core compiler modules to expose the spack-built
modules when a given compiler family is loaded.

In a future commit, the structure will be changed again to install the static
core modules in a tree which is not under the spack LMOD module root such that
`spack module lmod refresh --delete-tree` will only affect the spack-generated
modules and not remove the static modulefiles at all.

This commit is made now to preserve the working modulefile tree structure when
the static site modules are co-installed under the spack module root.