Syncs OMPI, SMPI, and Magma package forks with current production copies.
The OMPI, SMPI, and Magma packages used in the old SWCI framework (currently in
production as of 2020-06-18) have been updated since this spack-env-based
test repo was created. This commit pulls in the changes made that are currently
being used in production.

This commit also updates the notes of known spack issues and IBM bugs. Notably,
SMPI v10.3.1.2-20200121 added a function that is automatically inserted into
codes to return the SMPI version string. The declarations for this function are
broken under both GCC and XL and prevent netCDF4-c to correctly determine the
version of parallel-netcdf when building with MPI support.

In practice, once SMPI is installed, the `SMPI_PREFIX/include/mpi.h` file is
edited to remove this function by setting

```
if (1 == SMPI_HAVE_PRAGMA_WEAK)
```
to
```
if (2 == SMPI_HAVE_PRAGMA_WEAK)
```
completely disabling the problem block. We will need to check the next release
of SMPI to see if the problem declarations have been fixed.

It is also discovered that using Anaconda to provide a python3 environment for
spack to run exposes unwanted packages in the anaconda distro to the spack build
environment. We should strive to use either an OS-provided Python3 env or a
custom, minmial python virtualenv/conda env that does not include extra binaries
provided by Anaconda in the $PATH.