Removes extaneous specs and fixes package preferences for P9.
We don't need to test or build everything in this early stage. So we drop specs
for git-lfs, spectral, older HDF5 versions, and some of the tiered-spack
work-around specs in hopes of finding a better way to do mixed toolchain builds
when needed.

We also pin the netcdf-fortran version to one that works on P9. Building the
newer version causes some missing symbol errors at link time for some
toolchains. We will want to revisit this problem for netcdf-fortran in the near
future.

To reduce the number of unique HDF5 builds, we set preferences to build it with
support for the High-Level interface, C++ interface, and Fortran interface. This
configuration should be usable by more applications and save us from having 4
builds (with MPI support) of each HDF5 version for each compiler.