Splits general pacakge specs across toolchain classes.
Specs for large `defnitions:` sets of target applications that should be built
with each compiler should be expressed as a separate matrix for each compiler
class to more easily exclude problem builds and include more verbose exceptional
specs near the block in which packages are handled for a given class of
toolchain. IE:

```
- matrix:
  - - $general_compute_packages
  - - $gcc_compilers
  exclude:
  - 'amgx%gcc@9:'
- hpx %gcc@8.1.1 cxxstd=14 ^boost@1.70.0
- hpx %gcc@8.1.1 cxxstd=17 ^boost@1.70.0
- matrix:
  - - $general_compute_packages
  - - $llvm_compilers
  - - arch=linux-rhel7-power9le
  - - ^elfutils%gcc@4.8.5 arch=linux-rhel7-power8le
- matrix:
  - - $general_compute_packages
  - - $xl_compilers
  - - arch=linux-rhel7-power9le
  - - ^numactl%gcc@4.8.5 arch=linux-rhel7-power8le
  exclude:
  - boost@1.63.0:%xl
```
is better than:
```
- matrix:
  - - $general_compute_packages
  - - $all_compilers
  exclude:
  # lots of exceptions and separate matrixes for problem builds.
```

This way, specs for a given package that needs to be done differently for each
toolchain are not widely spread out across the manifest. This also allows
deployment of the whole stack for compiler classes that can do all the builds
reliably sooner than the tricky edge cases. Ie, we'll always have at least one
compiler environment that has everything.

This commit also adds exceptions for the mixed-toolchain builds involving XL
which cannot build perl due to some wierd issue with perls convoluted
multi-io-stream configure script being run under a python subprocess fork.

Similarly, exception builds for other toolchains are reorganized to be close to
the main matrix defining the bulk of the specs for each toolchain.