Adds WIP utility for removing orphaned specs from an environment.
When `spack concretize -f` is run, packages that were

- previously installed in an environment designated by path (and not a name living under `$spack/var/spack/environments/`)
- but no longer in the `spack.lock` manifest

become orphaned. They remain installed, but unknown to the base non-environment
spack instance which can only see the builds if the environment is activated.
However, if the environment is activated, spack won't uninstall the packages but
instead simply remove them from the environment.

This commit adds a utility that can be run under `spack-python` with an
activated environment to identify packages installed under the environment's
install_prefix but not in the environment manifest itself and actually uninstall
those packages.