Improves init script
Removes broken generic init script intended to initialize the environment for
arbitrary systems. Only the working but peak-specific init script remains.

Comments have been added to the peak-specific init script where it needs to be
generalized for arbitrary hosts.

The init script is modified to be aware of it's location on the fileystem so it
can be sourced from any CWD and now uses absolute path references relative to
it's location rather than relative paths to the CWD.

The spack instance has been moved from the host-specific directories to the root
directory. It should be possible with environments for all systems to share a
single spack instance. Lock files for operations on a specific environment
should exist within host/environment specific paths and prevent conflicts
arising from multiple systems using the same spack instance simultaneously.