Adds spack instance to repo as submodule.
This commit adds an OLCF-modified spack instance to the root of the
facility-spack-environments repo as a submodule. The source for this submodule
is a facility-controlled clone of vanilla spack at v0.14.2 with some
modifications to builtin packages. The changes to builtin packages are intended
to be generic and non-site-specific to allow us to more easily contribute
improvements to upstream spack as well as more easily pull changes down from upstream.

The submodule's `develop` branch tracks upstream spack. Changes to the spack
instance in the submodule are done in topic branches `olcf-X.Y.Z` that fork from
upstream spack at tagged releases `X.Y.Z`.

A specific commit to an OLCF-modified branch is tracked in the main
facility-spack-environments repo for cloning the submodule.