Corrects envvar configuration options
This commit updates the README with basic instructions for setting up a
development environment based on this repo.

Also adds a work-in-progress utility to show the internal values of
configuration options with expanded environment variables as spack would use
them to confirm that environment variables in the configuration files are being
used as expected.

And lastly it replaces as many hard-coded paths in the `spack.yaml` environment
configuration file with values inherited from environment variables. This should
help improve portability of the spack environments in private clones of this
repo.