Fixes script location self-identification logic.
Under Bash, the `$BASH_SOURCE` variable is always set, regardless if a bash
script is executed or sourced.

Under ZSH a bash script will only have `$BASH_SOURCE` set if the script is
executed and not if it is sourced.

Under both shells,

- `$0 != $BASH_SOURCE` when a bash script is sourced and
- `$0 == $BASH_SOURCE` when a bash script is executed.

However, when a bash script is sourced under bash, `$0` will be bash whereas
under ZSH it is the file that was sourced.

This commit adds an extra bit of logic to determine if, when a bash script is
sourced under either shell, if `$BASH_SOURCE` is set. If it is, we're in a bash
shell and can use the `BASH_SOURCE` var to determine the location of the script.
If it is not set, we are in a ZSH shell and can use the location of `$0` to
find the script being sourced.

It is not reliable to query `$SHELL` to determine which shell we are in as that
variable can be inproperly inherited across nested subshells.