Fixes patterns used to clean up common autofs paths.
The original script attempted to clean up the absolute autofs path for clones of
the environment under `/sw` but the pattern would mangle paths if the repo
existed under `/ccs/home`.

This commit replaces the autofs cleanup patterns with two that should work only
on `/sw` and `/ccs/home` but otherwise leave the path correct (but ugly under
autofs) if the repo were cloned on other filesytems.