Fixes install_tree root for expanded envvars.
Querying spack config for an environment's install_tree root does not
expand environment variables. However, querying the `spack.store`
database does. This change fixes an error where orphaned packages could
not be found because spack was looking in a non-existant install_tree
when the environment was configured to use environment variables to set
this path.