Updates OMPI and UCX package forks for Andes.
The UCX variant 'multithread' is renamed to 'thread_multiple' to match
the upstream OMPI spack package nomenclature.

Overly complex debug configuration is removed from the UCX package; we
want to trust the package's default self-configuration.

New versions of OMPI and UCX are added which are compatible with Andes
architecture and Mellanox network firmware.