Pins OMPI, UCX, RDMA-core versions for Mellanox driver.
The Mellanox driver packages installed on Andes provide some the
packages used by OMPI (parts of rdma-core, ucx, libfabrics, etc). It is
not generally possible to prevent builds from using these libraries and
headers. But it is also not possible to defer spack to use the
OS-provided versions of these packages to satisfy dependent specs.

Therefore, we must pin the versions of these packages in spack to what
is provided by the OS for compatibility with the mellanox firmware and
configure the builds to match/contain the subset of features provided by
the mellanox libraries in order to prevent missing symbol errors at
build time.

This commit also removes attempts to build netlib-scalapack 2.1.0
(latest available upstream) with GCC>=v10 due to a

"Rank mismatch in arguments error" for the F77 portions of the code.
This might be fixable by adjusting the `FCFLAGS` and `FC` toolchain used
by mpicc to better support older F77 codes under GCC10. Or the
netlib-scalapack package can be modified to add the necessary flags to
GCC10+ builds to overcome the issue. Worst case scenario is we'll have
to let an older GCC toolchain build this package for GCC10. However, in
the meantime, we are simply omitting these builds to carry on with
provisioning Andes.