Improves modulefile scheme.
This commit adds a whitelist exception for imagemagick anticipating that
we will either

- remove and supress implicit modules in the future
- or at least duplicate modules for packages that must be built with
variants of low-level dependencies.

In the later case, some packages must be built against ncurses with
termlib support (for LLVM optimization libraries) leading to duplicate
implicit modules for a number of common dependency packages (ncurses,
gobject-introspection, hdf5, etc).

This commit also enables TCL modulefile generation in anticipation that
we will offer a module that can expose all packages in a flat modulefile
tree. This would allow users to be able to conveniently (albeit with
great risk to the sanity of their build- and run-time environments) mix
packages from different toolchains or with different ABIs in their
environment.