Separates dependence on Core GCC for supporting PGI builds.
For packages that PGI cannot build, we use the OS GCC to support those
builds. However, to ensure that OMPI is not concretized different ways
for various packages (ie, all PGI builds of MPI-enabled packages use a
single build of OMPI), we must tightly control the PGI depenendency
specs built with GCC.

When using the Core GCC compiler spec, we end up with multiple builds of
core packages which results in modulefile name collisions. To avoid
this, we introduce a new compiler spec for the OS-provided GCC to be
used only when supporting PGI builds. Modulefiles for this compiler spec
are supressed.