Adds mechanism for switching between local/facility modulefile trees.
When we want to intialize the spack environment to build facility-provided
software using the facility's public default modulefiles, we don't need
to reset the `MODULEPATH` environment variable.

Only when the `ENV_PREFIX` doesn't match the facility software root do
we need to setup a custom "local" modulefile tree. This will be the case
when doing staging builds or relocating a copy of the environment for
private use by individual users.