Fixes matrix concretization error.
The excluded exception specs for dependents of netlib-scalapack built
with GCC10+ seem to have been causing an error where the whole set of
specs for GCC9+ were not being added to the environment.

This commit removes those exlusions in favor of more specific, explicit
exceptions.

This commit also re-organizes the GCC-only compute package specs to be
built after the core packages. The definition list is moved after the
core specs as well to match the order these specs will be built.