PGI fails to build hypre.
PGI is failing to build hypre. This spec is not needed among the
PGI-built packages because it available under the other toolchains. To
prevent this error failing the reset of the specs, it is excluded.
Ultimately, we will want to revisit this and ensure it's not a failure
that can be solved.