Updates spack to fix netcdf build error with GCC10+.
GCC10 fails to build netcdf-fortran and parallel-netcdf due to a
issue common to both packages that is known to the upstream developers
(at least for netcdf-fortran):

https://github.com/Unidata/netcdf-fortran/issues/212

NetCDF-fortran has fixed the issue in a new release as of early June 2020
(https://github.com/Unidata/netcdf-fortran/pull/250) but the problem has
not yet been patched in parallel-netcdf.

The upgrade to netcdf-fortran v4.5.3 requires an associated update to
netcdf-c.