Pins CUDA to v10 for kepler support.
Andes has kepler GPUs, support for which are becoming dropped and
deprecated in [CUDA11](https://docs.nvidia.com/cuda/cuda-toolkit-release-notes/index.html):

```
General CUDA
 - Support for Red Hat Enterprise Linux (RHEL) and CentOS 6.x is
 dropped.
- Support for Kepler sm_30 and sm_32 architecture based products is
  dropped.
- Support for the following compute capabilities are deprecated in
  the CUDA Toolkit:
  - sm_35 (Kepler)
  - sm_37 (Kepler)
  - sm_50 (Maxwell)
```

So for now we'll make CUDA10 the default, build everything that needs
CUDA support targetting kepler GPUs against it, but also provide CUDA11.
The nodes will get the CUDA11 firmware and drivers which should be
backwards/forwards compatible with the CUDA10 toolkit.

Gromacs and MAGMA will need to be patched to target only the
non-eliminated SM levels when building with CUDA11.