Fixes for spack upgrade to v0.15.0
- Spack v0.15.0 expands the granularity of compiler PIC flags to each
language but doesn't keep a backwards-compatible compiler class
attribute for generic PIC flags. The MPIP package is updated to use the
new flags attributes.

- The new builtin spack packages upgrade many dependencies. However,
the intel toolchain cannot build the latest version of CMake v3.17. This
appears to be a known issue/regression that has appeared in older
versions of CMake and will likely be patched in a future release. Until
then, we pin CMake to v3.16.2 for intel software builds.

- Additional specs added to the modulefile blacklists to permit a full
in-place refresh without name conflicts.