Updates README and refactors initialization process.
This commit updates the environment variables used during the spack
initialization process and in spack.yaml environment files to avoid
potential collisions with arbitrary variables in the environment by
adding variations of the prefix `FACSPACK` to any used by the init
script. The associated spack.yaml files and README are updated to use
the new variables.

Some of the variables are renamed to more clearly express their use.