Updates patch for PGI support for SMPI v10.4.0+
A new patch for PGI support was delivered by IBM to the OLCF on
2020-09-24 and is incorporated into SMPI installations, requiring a
change to the source tarball hash.

The patches provided by IBM are collected into a single tarball to
overwrite the contents of the flawed RPMs when the RPMs are unpackaged.
This patch tarball is included as a separate file in the SMPI "source"
tarball used by spack to define a current version. If the source tarball
includes a patch tarball, it will be used. Otherwise, the steps to patch
an RPM release of SMPI are skipped during the install phase.

Unfortunately, inclusion of the patch tarball in the source tarball
changes its checksum hash and requires the spack package to be updated.

The patching procedure is also updated to move the missing headers
provided by the patch to their standard locations when necessary.