Build OMPI with MLNX-provided UXC, RDMA-Core, and HCOLL on andes.
The MXM fabric is deprecated and replaced with the OFI fabric. HCOLL is
also available on Andes. So the default configuration for OMPI on Andes
is updated to use these fabrics. The Mellanox-provided dependencies to
support them are added as external path-driven packages.

Also, the module whitelist is cleaned up and only explicit packages are
given modulefiles to eliminate conflicts when regenerating the
modulefile tree during a period of OMPI test builds on Andes while
debugging IB network performance issues.

It may be necessary in the future to configure the OMPI runtime
environment to disable use of BTL TCP to avoid socket connection
timeouts. This can be done by adding

```
export OMPI_MCA_btl=^vader,openib,tcp
```

to the modulefiles for OMPI, for example.