Removes OMPI build from Peak
Issues finding liblsf/ls_info during the configuration stage are
preventing OMPI from being configured with LSF support on Peak. To
prevent this problem for an experimental package build from blocking
production software deployment, the OMPI specs are moved to a definition
block for special builds that is attempted after all standard production
software specs.

This special builds definition block is disabled until the OMPI
configuration errors against LSF can be fixed.

In the meantime, the vendor-provided path-based external packages for
Mellanox HCOLL and knem and IBM lsf are added to the package preferences
section of the environment for later use when OMPI builds can be
resumed.