Adds a mechanism to change build host in spack initialization script.
In some cases, notably during major OS upgrades, it is necessary to
build the software environment for a particular system from a node that
is running the upgraded node image. When this build node is not part of
the cannonical machine (eg, building a RHEL8 environment for Summit when
all Summit nodes are currently running RHEL7 by using a node from Peak
already running RHEL8), we need a mechanism to setup spack to build the
environment in a location that doesn't depend on information gathered
from the node itself.

This commit adds the ability to set the environment variable
`FACSPACK_HOST` different from the host detected by the init script.
Such that calling, for example,

```
FACSPACK_HOST="summit" . ./init-facility-spack.sh
```

from a Peak node will build the environment under `/sw/summit/...`.