Uses new spack-internal name for AOCC compilers.
Spack v0.16.0 now officially supports the AOCC toolchain as a unique
compiler class separate from the underlying LLVM/Clang implementation it
is derived from.

This commit replaces the compiler specs previously used for AOCC builds
on Andes with the new compiler spec name and updates the static
modulefiles for AOCC toolchains to expose the correct modulefile trees
when loaded.