Fixes package errors on Andes.
Andes OS compilers do not require a modulefile which, if requested but
not present, cause problems building harfbuzz due to modulecmd returning
unparsable json when setting up the build environment.

libtheora up to v1.1.1 needs libpng <1.6.0. This is expected to change
in the next release of libtheora. Package specs that depend on libtheora and
libpng (eg, paraview) must be constrained until the next release of libtheora.

Spack builtin packages for mpip and libtheora are updated with improved
dependency constraints.