Adds gcc@9.3.0, gcc@10.2.0, nvhpc@20.9 compilers to Summit, Peak
Adds modulefiles and specs for the following toolchains to both
Summit and Peak:

- GCC 9.3.0
- GCC 10.2.0
- Nvidia HPC SDK (nvhpc) 20.9

The SMPI spack package is updated to handle fortran mpi.mod for nvhpc.

Additionally, obsolete compiler specs are removed from the Peak base
software environment.

The nvhpc compilers also include a static modulefile for the CUDA
toolkit to preferentially use the in-built CUDA provided by nvhpc but
allow for users to opt for a different cuda toolkit if desired. The
nvhpc modulefiles include two for each version: one that exposes only
the runtime/math libs including cuda and one that also exposes the
compiler binaries.