Removes Arm-Forge installation on Andes
Arm Forge is a messy package to install and maintain via Spack. This
commit removes the installed copy of Arm Forge in favor of maintaining
Forge manually outside of Spack.

This allows us the following UX improvements:

- We can name the modulefiles simply "forge", consistant with our other
  machines
- The version identifiers are simple eg "20.1" instead of
  "20.1-Redhat-8.0-x86_64". Lmod global aliases and package alias don't
  present the aliases in a user-friendly way or allow for hiding the
  required convoluted name used by Spack if Forge were installed that
  way.
- There is no risk of leaving `allinearc` blank after spack environment
  rebuilds.

Plus Forge is distributed as a binary package so it's easy to install.
The installation prefix has been moved to `/sw/andes/forge/X.Y` with
manually written modulefiles installed to
`/sw/andes/modulefiles/core/forge`.

Despite this, we will keep the spackage fork so it is possible to fetch
the source for manual installations using `spack stage arm-forge@...`.