Upgrades SMPI to v10.4.0.3-20210112
The latest version of SMPI is added to the spack package and Peak
environment. The spack package is refactored to simplify the logic used
in post-install steps that changes for different compilers each new
release of SMPI.

Common tasks of

- accepting the license
- patching a bug out of `mpi.h` which must be done post-install or
  within the install method due to the way we're installing SMPI from
  RPMs that precludes the use of spack's patch methods.

are sequentially ordered in a single `run_after('install')` method
followed by a logic switch that applies version-specific tasks, such as
rebuilding the mpi-fortran modules and one-off fixes that don't
generalize well between release versions.