Fixes path to XL 16.1.1-10 in modulefiles for Summit.
A common XL PTF10 install can be used for both RHEL7 and RHEL8.
Therefore, the alternate install 16.1.1-10-1 intended for RHEL8 was
never installed. However, the modulefile for XL PTF10 was written to use
that path prior to XL actually being installed. This commit corrects the
malformed path to the XL installation when used on Summit (RHEL8).