Adds Spock awareness to spack init script.
The first OLCF5 EA system `spock.olcf.ornl.gov` uses a different
hostname scheme than previous machines. An exception is added to
positively ID the host when working on Spock builds.

When a system does not have static modulefiles tracked in this repo, an
exception is added to skip attempting to copy modulefiles from a
non-existent or empty directory for that host.

Spock's modulefile system is a work-in-progress and is, at the time of
this writing, generally broken for ZSH users. When setting up spack's
runtime enviroment $MODULEPATH, we must currently reconstruct the
variable from the ground up and ensure spack-built module directories
are added to the search list.