Separates ROCm specs into a separate spec definition list.
Due to the number of ROCm specs and the need to support multiple
versions, we express the ROCm software stack in terms of a definition
list to allow building them in preferred order and with different
versions as needed.

The list of allowed compilers is also reduced to match what is currently
usable on the system.