Adds initial E4S packages to Spock
Uses a fixed, facility-wide source cache when possible.

Multiple instances of this repo are needed simultaneously at different
commits to handle parallel builds of environments for machines requiring
different versions of the spack builtin repository. To reduce the amount
of storage dedicated to cached package source archives, each environment
is updated to use a common shared cache if it exists and is writable by
the user performing software builds.

Increases build parallelism on spock.

Makes improvements to E4S packages on Cray Shasta.