Adds AOCC 3.0.0 specs to Andes
AOCC 3.0.0 fails to build HDF5 due to flang compiler driver interface
errors documented in
https://jira.ccs.ornl.gov/projects/OLCFSW/queues/custom/131/OLCFSW-71

Builds for that toolchain have been separated from the common matrix of
aocc build specs in an attempt to both exclude fewer general-compute
packages and to see if HDF5 v1.12.0 would have fixes to allow flang to
be used as the fortran language compiler.  However, the problems persist
in HDF5 v1.12.0.