Suppresses unwanted implicit modulefiles on Spock
For now, we currently want to keep modulefile generation for implicit
packages due to an elusive bug in spack where blacklisting all implicit
packages sometimes blocks LMOD module generation for all packages,
including those that are explicitly installed.

However, we want to reduce the clutter and overwhelming number of
unecessary modulefiles in order to give users a better experience and
make useful target packages more visible in environment module views.

Therefore, we explicitly blacklist many of the common implicitly
installed packages that users will unlikely want or independently need.
Individual packages can be un-blacklisted if requested (and, if
requested, should also be marked explicit and added to the environment
spec list).

Most notably, the individual ROCm components are blacklisted; we want to
expose ROCm as a complete stack through a single view and metamodule on
each ROCm version release.