Removes Score-P spec from Spock until failure is fixed upstream.
The latest Score-P has dependencies on CubeW and and Cube-Lib which, at
the current latest version, have hard-coded mechanisms in their
autotools build system to determine the target Cray platform that do not
accomodate Shasta systems. Until this issue is fixed upstream or patched
locally, we cannot build these dependencies.