Updates ROCm view generation scripts
A script to generate a ROCm view for ROCm v4.2.0 is added. To make
altering specs for individual packages in the views, the structure of
the spec list is changed to allow line-item modification as well as spec omission via line comments.

The view generation scripts have also been made more consistant between
each version in anticipation of eventually merging these scripts into a
single script that can handle multiple ROCm versions.

A generic script to find all the specs in an environment associated with
a ROCm deployment `find-rocm-specs.sh` has been added to aid in locating
duplicate component builds and identifying specific build hashes for
ROCm views that must contend with duplicate components installed in the
environment. This is notably a problem due to

- rocm-openmp-extras being separate from llvm-amdgpu which causes
  multiple builds of rocm-openmp-extras dependencies due to differences
  in variants on llvm-amdgpu required by different parts of a ROCm stack
- A circular dependency between
  `rocm-openmp-extras-->(hsa-rocr-dev, rocm-device-libs) --> llvm-amdgpu --> rocm-openmp-extras`
  which must be resolved at either the ROCm component source level or a
  fundamental restructuring of the ROCm spack packages.
  It is necessary for rocm-openmp-extras to be merged into llvm-amdgpu
  but if it depends on staged/intermediate builds of hsa-rocr-dev and
  rocm-device-libs requiring llvm-amdgpu, then they should all be merged
  into a single package. Or an additional llvm-amdgpu-bootstrap package
  (or other suitable LLVM toolchain) should exist which won't be
  considered as part of a ROCm release to be used to break the circular
  dependency.

Lastly, the view projection files for each ROCm release are made to be
symlinks pointing to a generic projection file. The projection file is
needed to form a spack view of an entire ROCm stack under a common
prefix. A common-prefix view is necessary because many of the ROCM
components locate their dependencies at runtime relative to themselves
or relative to a set of semi-standardized (though inconsistently used
among the ROCm components) environment variables like `$ROCM_PATH`,
`$HIP_PATH`, `$LLVM_PATH`, etc which are not useful to form relative
paths to other ROCm components when each are installed in separate
prefixes under spack.

Version-specific symlinks to a generic projection file are used to allow
the option to change the projection per-version for future ROCm releases
should that be necessary while minimizing the number of projection files
and code duplication that we must maintain.