Moves GCC-only spec definition list directly into GCC spec list
There's no need to have a short definition list of specs to be built
only with GCC separate from the definition list for all other GCC-built
specs. Also, raja and papi are included in the E4S spec list and must be
excluded separately for each toolchain that is unable to build them
anyway.

This commit also supresses builds of E4S packages by GCC10+ that are
incompatible with that toolchain at their currently available versions.