Suppress conflicting scalapack modulefiles for refresh
Multiple builds of netlib-scalapack dependent on openblas exist on the system
for various versions of openblas that would suffer a modulefile name collision
when the complete set of generated modulefiles are refreshed. We want only the
builds that depend on the latest openblas to be generated. This commit
suppresses modulefile generation for the unwanted scalapack builds.