Updates ROCm view generation scripts.
Fixes empty dirs in view for directories in ROCm component packages that
are relative symlinks to self or other ROCm components. In the case
where the symlink in the component's install prefix references itself
(ie the install prefix is intended to be the ROCM_PATH), the symlinks
are cast into a spack view as empty directories. In the case where
relative symlinks reference other ROCm componments, the links point to
the wrong location when spack installs the component in it's own private
prefix.

Adds missing ROCm component `rocm-dbgapi` to views.