Adds PE 21.07 components and compilers, pin specs to PE 21.04 dependencies
This commit adds the PE21.07 compilers
- cce 12.0.1
- cce 12.0.0
- gcc 10.3.0

along with external specs for

- ROCm v4.2.0
- cray-mpich 8.1.6 and 8.1.7
- cray-libsci v21.06.1.1

but pins all builds using PE 21.04 compilers against PE 21.04
dependencies (`cray-mpich@8.1.4`, `cray-libsci@21.04.1.1`,
`/opt/rocm-4.1.0`).

cce@12.0.0 from PE 21.06 specs are added to use PE21.06 dependencies
with the exception that libsci or openblas provider partial specs must
be added to each spec in a spec matrix. Otherwise, the global defaults
pinned at the `packages:all:{blas,lapack}` global preferences will be
used. Here, we've pinned the order of libsci preference to be in the
order

[Generic OpenBLAS, PE21.04, PE21.06]

to avoid un-necessarily re-building previous successful builds agaisnt
the latest PE libsci. But this isn't reliable since the PE21.06
compilers also use PE21.04 libsci when libsci is preferred over
openblas. We need to explicitly add blas/lapack providers to each spec
in order to be able to use different versions of libsci for different
PE-release compilers but can't add a blanket matrix element for libsci
to spec matrixes since we sometimes want to use openblas. This is
a miserable situation and means we either cannot rely on spec matrixes
or must split e4s spec matrixes into separate ones for each blas/lapack
provider between (openblas, libsci). Unless we can use
`packages:<package>:providers:{blas,lapack}` to steer spack to use
libsci vs openblas with different versions of libsci matching the PE
level of the compiler used to build the spec. It is not clear if this is
possible.