Removes craype-accel-amd-gfx908 module from cce compiler pre-requisite modules
There are undefined symbols in the cray-mpich GTL libraries which are
linked into all MPI applications by the cray wrappers when the
craype-accel-amd-gfx908 module is loaded. The GTL libraries need AMD's
libhsa_runtime library linked into applications to resolve these
symbols but CrayPE does not provide the HSA runtime. Because the HSA
runtime library is installed to a potentially arbitrary location when
multiple versions of ROCm are available either under /opt/rocm-X.Y.Z
or in an alternate prefix, the GTL libs do not provide RUNPATH or RPATHS
to a fixed HSA runtime and the cray wrappers do not put the HSA runtime
into the compiler invocation or otherwise add them to the LIBRARY_PATH
and LD_LIBRARY_PATH at both compile time and runtime. While this is a
problem at runtime for applications that depend on MPI if these symbols
are not stripped from binaries if/when they are not used, a more
pressing complication is that CCE treats undefined symbols in shared
libraries as a hard fatal error and cannot compile virtually any
application that does also explictly depend on ROCm and provide link
arguments to a suitable HSA runtime library and ensure it is available
to the binary at later runtime (which is often not the case for
configure-time and compile-time tests).

To resolve these issues, we no longer load the craype-accel-amd-gfx908
modules with the CCE compiler spec such that the cray wrappers are not
generally configured to link in the cray-mpich GTL libs.

To test this, we also re-enable previously exluded CCE build specs that
possibly failed due to this cray compiler wrapper nonsense and broken
craype-provided libraries.