Adds GCC 11.1.0 builds to Summit RH8
Adds GCC 11.1.0 builds to Summit consisting of as many applicable E4S
21.05 (latest) release packages as possible.

However, since the CUDA toolkit up through the current release 11.4.0
does not officially support GCC 11 as a host compiler, we have excluded
most packages which need nvcc at build time. It is probably best to not
build these packages until a suitable CUDA toolkit is available in order
to not give the illusion that they are optimized for Summit hardware.
Only at user request should we add packages currently skipped if built
without CUDA support:

- arborx+cuda
- cabana+cuda
- chai+cuda
- hpctoolkit+cuda
- kokkos-kernels+cuda
- kokkos+cuda
- legion+cuda
- magma+cuda
- papi+cuda
- petsc+cuda
- raja+cuda
- strumpack+cuda
- sundials+cuda
- superlu-dist+cuda
- slate+cuda
- tasmanian+cuda
- tau+cuda
- umpire+cuda
- upcxx+cuda

Some packages have been identified as optionally dependent on CUDA and
we ought to make them CUDA-enabled by default for all builds with all
compilers:

- amrex+cuda
- axom+cuda
- caliper+cuda
- ginkgo+cuda
- heffte+cuda
- hpx+cuda

with the one exception of `zfp+cuda`; a compression algorithm package
that, as a common dependency, might end up in builds targetted for the
login nodes.

Of the above optionally-cuda-enabled packages, it was found that our
current build defaults prefer a cuda-disabled variant of the target
package but a cuda-enabled variant of a dependency. These builds are
skipped for GCC 11 as stated above, our package preferences should be
updated to prefer the cuda-enabled root spec:

- axom~cuda ^raja+cuda
- caliper~cuda ^papi+cuda
- precice ^superlu-dist~cuda
- slepc ^superlu-dist~cuda