Updates modulefile inspections and blacklist for Summit
Modulefile suffixes are added to builds for which we want to keep
modules that would otherwise have name conflicts (kokkos, petsc, and
umpire).

To solve a testshot issue and make the modulefile contents consistent
with Summit RH7, we have re-enabled `CPATH` and `LIBRARY_PATH`
inspections for all modules.

And finally, a number of superfluous modules are added to the blacklist.
Some of these are benign and are only suppressed to reduce the
overwhelming number of modules that users will see when querying the
module system. Others prevent in-place modulefile tree regeneration due
to multiple builds causing modulefile name conflicts.

With the blacklist and suffixes updated, it is possible to
interactively regenerate the modulefile tree to expose the latest builds
of all packages on Summit for it's return to users after the upgrade
outage on 2021-08-17 through 2021-08-19.