Removes temporary cray-mpich dependent module blacklist specs
Due to the way libsci is dependent on cray-mpich, some packages that are
implicitly MPI-dependent through libsci are not given modules that are
located under MPI-version specific module directories. These modules
exist in the "core" modules dir.

We need to either

- restructure the module tree to have a cray-libsci layer in the
  heirarchy
- blacklist "core" modules built against previous libsci/mpich verions
  that conflict with new builds when refreshing modules.
- or include only specs in the manifest that use a single version of
  libsci/mpich

In this case, we've opted to blacklist the conflicts but should probably
just build specs against a single version of libsci/mpich so conflicts
don't arise again in the future. Adding a libsci layer to the module
heirarchy will introduce enormous complexity and is likely not worth the
effort. Unfortunately, this means that we will effectively support
builds linked against only one version of CPE at a time.